+++ extras/AFC_buffer.py
@@ -13,15 +13,17 @@ class AFCtrigger:
 
     def __init__(self, config):
         self.printer = config.get_printer()
-        self.reactor = self.printer.get_reactor()
-        self.gcode = self.printer.lookup_object('gcode')
+        self.AFC = self.printer.lookup_object('AFC')
+        self.reactor = self.AFC.reactor
+        self.gcode = self.AFC.gcode
+
         self.name = config.get_name().split(' ')[-1]
         self.turtleneck = False
         self.belay = False
         self.last_state = False
         self.enable = False
         self.current = ''
-        self.AFC = self.printer.lookup_object('AFC')
+
         self.debug = config.getboolean("debug", False)
         self.buttons = self.printer.load_object(config, "buttons")
 
@@ -66,7 +68,6 @@ class AFCtrigger:
             raise error( msg )
 
         self.printer.register_event_handler("klippy:ready", self._handle_ready)
-
         self.gcode.register_mux_command("QUERY_BUFFER", "BUFFER", self.name, self.cmd_QUERY_BUFFER, desc=self.cmd_QUERY_BUFFER_help)
 
         # Belay Buffer
@@ -160,27 +161,23 @@ class AFCtrigger:
     def advance_callback(self, eventime, state):
         if self.printer.state_message == 'Printer is ready' and self.enable:
             CUR_LANE = self.printer.lookup_object('AFC_stepper ' + self.AFC.current)
-            CUR_EXTRUDER = self.printer.lookup_object('AFC_extruder ' + CUR_LANE.extruder_name)
-            if CUR_EXTRUDER.tool_start_state:
-                if self.AFC.current != None and state:
-                    CUR_LANE.assist(CUR_LANE.calculate_pwm_value(self.AFC.gcode_move.speed * (self.velocity / 10)))
-                    self.reactor.pause(self.reactor.monotonic() + 1)
-                    CUR_LANE.assist(0)
-                    self.set_multiplier( self.multiplier_low )
-                    if self.debug: self.gcode.respond_info("Buffer Triggered State: Advanced")
+            if self.AFC.current != None and state:
+                CUR_LANE.assist(CUR_LANE.calculate_pwm_value(self.AFC.gcode_move.speed * (self.velocity / 10)))
+                self.reactor.pause(self.reactor.monotonic() + 1)
+                CUR_LANE.assist(0)
+                self.set_multiplier( self.multiplier_low )
+                if self.debug: self.gcode.respond_info("Buffer Triggered State: Advanced")
         self.last_state = ADVANCE_STATE_NAME
 
     def trailing_callback(self, eventime, state):
         if self.printer.state_message == 'Printer is ready' and self.enable:
             CUR_LANE = self.printer.lookup_object('AFC_stepper ' + self.AFC.current)
-            CUR_EXTRUDER = self.printer.lookup_object('AFC_extruder ' + CUR_LANE.extruder_name)
-            if CUR_EXTRUDER.tool_start_state:
-                if self.AFC.current != None and state:
-                    CUR_LANE.assist(CUR_LANE.calculate_pwm_value(self.AFC.gcode_move.speed * (self.velocity / 10)))
-                    self.reactor.pause(self.reactor.monotonic() + 1)
-                    CUR_LANE.assist(0)
-                    self.set_multiplier( self.multiplier_high )
-                    if self.debug: self.gcode.respond_info("Buffer Triggered State: Trailing")
+            if self.AFC.current != None and state:
+                CUR_LANE.assist(CUR_LANE.calculate_pwm_value(self.AFC.gcode_move.speed * (self.velocity / 10)))
+                self.reactor.pause(self.reactor.monotonic() + 1)
+                CUR_LANE.assist(0)
+                self.set_multiplier( self.multiplier_high )
+                if self.debug: self.gcode.respond_info("Buffer Triggered State: Trailing")
         self.last_state = TRAILING_STATE_NAME
 
     def buffer_status(self):